refactor(0081): split grid_db/repositories.py into a package (#184)
First incremental extraction for issue #184 (P2 module hygiene). Convert the
1512-line flat repositories.py into a grid_db/repositories/ package of
domain-grouped modules — base (BaseRepository + T), identity, market_data,
execution, snapshots — with __init__.py re-exporting all 13 classes.

ZERO behavior change: class bodies, signatures, and docstrings are byte-for-byte
identical (verified); only import headers and file boundaries differ. Both
`from grid_db import XRepository` and `from grid_db.repositories import XRepository`
keep resolving (grid_db/__init__.py and all callers unchanged).

- Add shared/db/tests/test_repository_imports.py — dual-path identity guard over
  all 13 classes + BaseRepository.
- Update RULES.md with the repository-module layout + re-export rule.

Plan: docs/features/0081_PLAN.md (adversarial plan-debate, approved iter 1).
